Changes to enable rate to be < 1.

+++ src/main/java/com/oltpbenchmark/WorkloadConfiguration.java
@@ -148,7 +148,7 @@ public class WorkloadConfiguration {
         this.workloadState = new WorkloadState(benchmarkState, phases, terminals);
     }
 
-    public void addPhase(int id, int time, int warmup, int rate, List<Double> weights, boolean rateLimited, boolean disabled, boolean serial, boolean timed, int active_terminals, Phase.Arrival arrival) {
+    public void addPhase(int id, int time, int warmup, double rate, List<Double> weights, boolean rateLimited, boolean disabled, boolean serial, boolean timed, int active_terminals, Phase.Arrival arrival) {
         phases.add(new Phase(benchmarkName, id, time, warmup, rate, weights, rateLimited, disabled, serial, timed, active_terminals, arrival));
     }
